+++ src/main/java/domain/result/Income.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 public class Income {
